Refactor intra view event dispatcher
Renamed from client-events to intra-view-events to be more specific.
Renamed from listener to dispatcher to be more specific.

Remove remote storage (GCM) functionality out of the intra-view-events
dispatcher.

Moved addition of handlers into the DI area (start.js) which reduces
dependencies in the dispatcher from 5 down to 2 and simplifies testing.

Added a test to confirm correct dispatching

Updated tests to return promises to mocha
